{"schema":"decodex/commit/1","summary":"Move backend window list runtime","authority":"manual"}
